Added CAxisAlignedBoundingBox for testing against frustum for culling. Got marching cubes looking good just has a few problems to sort out. Almost finished cloth simulation.

+++ Slimfish/Source/Graphics/SlimVertexDeclaration.cpp
@@ -156,7 +156,7 @@ const CInputElement& CVertexDeclaration::GetElement(size_t index) const
 	return m_ElementList[index];
 }
 
-int CVertexDeclaration::GetNumElements() const
+size_t CVertexDeclaration::GetNumElements() const
 {
 	return m_ElementList.size();
 }
